+++ tests/integration-tests/src/test/java/org/apache/activemq/artemis/tests/integration/ra/OutgoingConnectionNoJTATest.java
@@ -17,20 +17,6 @@
 
 package org.apache.activemq.artemis.tests.integration.ra;
 
-import javax.jms.Connection;
-import javax.jms.JMSContext;
-import javax.jms.JMSException;
-import javax.jms.JMSProducer;
-import javax.jms.Message;
-import javax.jms.MessageConsumer;
-import javax.jms.MessageProducer;
-import javax.jms.Queue;
-import javax.jms.QueueConnection;
-import javax.jms.Session;
-import javax.jms.TextMessage;
-import java.util.HashSet;
-import java.util.Set;
-
 import org.apache.activemq.artemis.api.core.client.ClientConsumer;
 import org.apache.activemq.artemis.api.core.client.ClientMessage;
 import org.apache.activemq.artemis.api.core.client.ClientSession;
@@ -49,7 +35,20 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
+import javax.jms.Connection;
+import javax.jms.JMSContext;
+import javax.jms.JMSProducer;
+import javax.jms.Message;
+import javax.jms.MessageConsumer;
+import javax.jms.MessageProducer;
+import javax.jms.Queue;
+import javax.jms.QueueConnection;
+import javax.jms.Session;
+import javax.jms.TextMessage;
+import java.util.HashSet;
+import java.util.Set;
+
+public class OutgoingConnectionNoJTATest extends ActiveMQRATestBase {
 
    protected ActiveMQResourceAdapter resourceAdapter;
    protected ActiveMQRAConnectionFactory qraConnectionFactory;
@@ -71,18 +70,18 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
       ((ActiveMQJAASSecurityManager) server.getSecurityManager()).getConfiguration().setDefaultUser("guest");
       ((ActiveMQJAASSecurityManager) server.getSecurityManager()).getConfiguration().addRole("testuser", "arole");
       ((ActiveMQJAASSecurityManager) server.getSecurityManager()).getConfiguration().addRole("guest", "arole");
-      Role role = new Role("arole", true, true, true, true, true, true, true, true, true, true);
+      Role role = new Role("arole", true, true, true, true, true, true, true, true);
       Set<Role> roles = new HashSet<>();
       roles.add(role);
       server.getSecurityRepository().addMatch(MDBQUEUEPREFIXED, roles);
 
       resourceAdapter = new ActiveMQResourceAdapter();
       resourceAdapter.setEntries("[\"java://jmsXA\"]");
-
       resourceAdapter.setConnectorClassName(InVMConnectorFactory.class.getName());
       MyBootstrapContext ctx = new MyBootstrapContext();
       resourceAdapter.start(ctx);
       mcf = new ActiveMQRAManagedConnectionFactory();
+      mcf.setAllowLocalTransactions(true);
       mcf.setResourceAdapter(resourceAdapter);
       qraConnectionFactory = new ActiveMQRAConnectionFactoryImpl(mcf, qraConnectionManager);
    }
@@ -100,17 +99,17 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
    }
 
    @Test
-   public void testSimpleMessageSendAndReceiveTransacted() throws Exception {
-      setDummyTX();
+   public void testSimpleMessageSendAndReceiveSessionTransacted() throws Exception {
       setupDLQ(10);
       resourceAdapter = newResourceAdapter();
       MyBootstrapContext ctx = new MyBootstrapContext();
       resourceAdapter.start(ctx);
       ActiveMQRAManagedConnectionFactory mcf = new ActiveMQRAManagedConnectionFactory();
+      mcf.setAllowLocalTransactions(true);
       mcf.setResourceAdapter(resourceAdapter);
       ActiveMQRAConnectionFactory qraConnectionFactory = new ActiveMQRAConnectionFactoryImpl(mcf, qraConnectionManager);
       QueueConnection queueConnection = qraConnectionFactory.createQueueConnection();
-      Session s = queueConnection.createSession(true, Session.AUTO_ACKNOWLEDGE);
+      Session s = queueConnection.createSession(true, Session.SESSION_TRANSACTED);
       Queue q = ActiveMQJMSClient.createQueue(MDBQUEUE);
       MessageProducer mp = s.createProducer(q);
       MessageConsumer consumer = s.createConsumer(q);
@@ -126,60 +125,90 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
       assertNotNull(textMessage);
       assertEquals(textMessage.getText(), "test");
       s.commit();
+      textMessage = (TextMessage) consumer.receiveNoWait();
+      assertNull(textMessage);
    }
 
-   public void testQueuSessionAckMode(boolean inTx) throws Exception {
-      if (inTx) {
-         setDummyTX();
-      }
+   @Test
+   public void testSimpleMessageSendAndReceiveNotTransacted() throws Exception {
+      setupDLQ(10);
+      resourceAdapter = newResourceAdapter();
+      MyBootstrapContext ctx = new MyBootstrapContext();
+      resourceAdapter.start(ctx);
+      ActiveMQRAManagedConnectionFactory mcf = new ActiveMQRAManagedConnectionFactory();
+      mcf.setAllowLocalTransactions(true);
+      mcf.setResourceAdapter(resourceAdapter);
+      ActiveMQRAConnectionFactory qraConnectionFactory = new ActiveMQRAConnectionFactoryImpl(mcf, qraConnectionManager);
+      QueueConnection queueConnection = qraConnectionFactory.createQueueConnection();
+      Session s = queueConnection.createSession(false, Session.SESSION_TRANSACTED);
+      Queue q = ActiveMQJMSClient.createQueue(MDBQUEUE);
+      MessageProducer mp = s.createProducer(q);
+      MessageConsumer consumer = s.createConsumer(q);
+      Message message = s.createTextMessage("test");
+      mp.send(message);
+      s.commit();
+      queueConnection.start();
+      TextMessage textMessage = (TextMessage) consumer.receive(1000);
+      assertNotNull(textMessage);
+      assertEquals(textMessage.getText(), "test");
+      s.rollback();
+      textMessage = (TextMessage) consumer.receive(1000);
+      assertNotNull(textMessage);
+      assertEquals(textMessage.getText(), "test");
+      s.commit();
+      textMessage = (TextMessage) consumer.receiveNoWait();
+      assertNull(textMessage);
+   }
+
+   @Test
+   public void testSimpleMessageSendAndReceiveSessionTransacted2() throws Exception {
+      setupDLQ(10);
+      resourceAdapter = newResourceAdapter();
+      MyBootstrapContext ctx = new MyBootstrapContext();
+      resourceAdapter.start(ctx);
+      ActiveMQRAManagedConnectionFactory mcf = new ActiveMQRAManagedConnectionFactory();
+      mcf.setAllowLocalTransactions(true);
+      mcf.setResourceAdapter(resourceAdapter);
+      ActiveMQRAConnectionFactory qraConnectionFactory = new ActiveMQRAConnectionFactoryImpl(mcf, qraConnectionManager);
+      QueueConnection queueConnection = qraConnectionFactory.createQueueConnection();
+      Session s = queueConnection.createSession(Session.SESSION_TRANSACTED);
+      Queue q = ActiveMQJMSClient.createQueue(MDBQUEUE);
+      MessageProducer mp = s.createProducer(q);
+      MessageConsumer consumer = s.createConsumer(q);
+      Message message = s.createTextMessage("test");
+      mp.send(message);
+      s.commit();
+      queueConnection.start();
+      TextMessage textMessage = (TextMessage) consumer.receive(1000);
+      assertNotNull(textMessage);
+      assertEquals(textMessage.getText(), "test");
+      s.rollback();
+      textMessage = (TextMessage) consumer.receive(1000);
+      assertNotNull(textMessage);
+      assertEquals(textMessage.getText(), "test");
+      s.commit();
+      textMessage = (TextMessage) consumer.receiveNoWait();
+      assertNull(textMessage);
+   }
+
+   @Test
+   public void sessionTransactedTestNoActiveJTATx() throws Exception {
+      JMSContext context = qraConnectionFactory.createContext(JMSContext.SESSION_TRANSACTED);
+      assertEquals(context.getSessionMode(), JMSContext.AUTO_ACKNOWLEDGE);
+   }
+
+
+   @Test
+   public void testQueuSessionAckMode() throws Exception {
+
       QueueConnection queueConnection = qraConnectionFactory.createQueueConnection();
 
-      Session s = queueConnection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-      if (inTx) {
-         assertEquals(Session.SESSION_TRANSACTED, s.getAcknowledgeMode());
-      } else {
-         assertEquals(Session.AUTO_ACKNOWLEDGE, s.getAcknowledgeMode());
-      }
+      Session s = queueConnection.createSession(false, Session.SESSION_TRANSACTED);
+
       s.close();
-
-      s = queueConnection.createSession(false, Session.DUPS_OK_ACKNOWLEDGE);
-      if (inTx) {
-         assertEquals(Session.SESSION_TRANSACTED, s.getAcknowledgeMode());
-      } else {
-         assertEquals(Session.DUPS_OK_ACKNOWLEDGE, s.getAcknowledgeMode());
-      }
-      s.close();
-
-      //exception should be thrown if ack mode is SESSION_TRANSACTED or
-      //CLIENT_ACKNOWLEDGE when in a JTA else ackmode should bee ignored
-      try {
-         s = queueConnection.createSession(false, Session.SESSION_TRANSACTED);
-         if (inTx) {
-            assertEquals(s.getAcknowledgeMode(), Session.SESSION_TRANSACTED);
-         } else {
-            fail("didn't get expected exception creating session with SESSION_TRANSACTED mode ");
-         }
-         s.close();
-      } catch (JMSException e) {
-         if (inTx) {
-            fail("shouldn't throw exception " + e);
-         }
-      }
-
-      try {
-         s = queueConnection.createSession(false, Session.CLIENT_ACKNOWLEDGE);
-         if (inTx) {
-            assertEquals(s.getAcknowledgeMode(), Session.SESSION_TRANSACTED);
-         } else {
-            fail("didn't get expected exception creating session with CLIENT_ACKNOWLEDGE mode");
-         }
-      } catch (JMSException e) {
-         if (inTx) {
-            fail("shouldn't throw exception " + e);
-         }
-      }
-
    }
+
+
 
    @Test
    public void testSimpleSendNoXAJMSContext() throws Exception {
@@ -187,7 +216,7 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
 
       try (ClientSessionFactory sf = locator.createSessionFactory();
            ClientSession session = sf.createSession();
-           ClientConsumer consVerify = session.createConsumer(MDBQUEUE);
+           ClientConsumer consVerify = session.createConsumer("jms.queue." + MDBQUEUE);
            JMSContext jmsctx = qraConnectionFactory.createContext();
       ) {
          session.start();
@@ -201,16 +230,6 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
          assertNotNull(msg);
          assertEquals("hello", msg.getStringProperty("strvalue"));
       }
-   }
-
-   @Test
-   public void testQueueSessionAckModeJTA() throws Exception {
-      testQueuSessionAckMode(true);
-   }
-
-   @Test
-   public void testSessionAckModeNoJTA() throws Exception {
-      testQueuSessionAckMode(false);
    }
 
    @Test
@@ -233,7 +252,7 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
       Queue q = ActiveMQJMSClient.createQueue(MDBQUEUE);
       try (ClientSessionFactory sf = locator.createSessionFactory();
            ClientSession session = sf.createSession();
-           ClientConsumer consVerify = session.createConsumer(MDBQUEUE);
+           ClientConsumer consVerify = session.createConsumer("jms.queue." + MDBQUEUE);
            Connection conn = qraConnectionFactory.createConnection();
       ) {
          Session jmsSess = conn.createSession(false, Session.AUTO_ACKNOWLEDGE);
@@ -248,9 +267,5 @@ public class OutgoingConnectionTestJTA extends ActiveMQRATestBase {
          assertNotNull(msg);
          assertEquals("hello", msg.getStringProperty("strvalue"));
       }
-   }
-
-   private void setDummyTX() {
-      ((DummyTransactionManager) ServiceUtils.getTransactionManager()).tx = new DummyTransaction();
    }
 }
