+++ tests/integration-tests/src/test/java/org/apache/activemq/artemis/tests/integration/stomp/StompWithMessageIDTest.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements. See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,10 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.activemq.artemis.tests.integration.stomp;
 
+import javax.jms.Message;
 import javax.jms.MessageConsumer;
+import javax.jms.QueueBrowser;
 import javax.jms.TextMessage;
+import java.util.Enumeration;
 
 import org.apache.activemq.artemis.tests.integration.stomp.util.ClientStompFrame;
 import org.apache.activemq.artemis.tests.integration.stomp.util.StompClientConnection;
@@ -25,40 +29,52 @@ import org.apache.activemq.artemis.tests.integration.stomp.util.StompClientConne
 import org.junit.Assert;
 import org.junit.Test;
 
-public class StompTestWithSecurity extends StompTestBase {
+public class StompWithMessageIDTest extends StompTestBase {
 
    @Override
-   public boolean isSecurityEnabled() {
+   public boolean isEnableStompMessageId() {
       return true;
    }
 
    @Test
-   public void testJMSXUserID() throws Exception {
-      server.getActiveMQServer().getConfiguration().setPopulateValidatedUser(true);
-
-      MessageConsumer consumer = session.createConsumer(queue);
-
+   public void testEnableMessageID() throws Exception {
       StompClientConnection conn = StompClientConnectionFactory.createClientConnection(uri);
       conn.connect(defUser, defPass);
 
       ClientStompFrame frame = conn.createFrame("SEND");
       frame.addHeader("destination", getQueuePrefix() + getQueueName());
-      frame.setBody("Hello World");
+      frame.setBody("Hello World 1");
       conn.sendFrame(frame);
 
-      conn.disconnect();
+      frame = conn.createFrame("SEND");
+      frame.addHeader("destination", getQueuePrefix() + getQueueName());
+      frame.setBody("Hello World 2");
+      conn.sendFrame(frame);
+
+      QueueBrowser browser = session.createBrowser(queue);
+
+      Enumeration enu = browser.getEnumeration();
+
+      while (enu.hasMoreElements()) {
+         Message msg = (Message) enu.nextElement();
+         String msgId = msg.getStringProperty("amqMessageId");
+         assertNotNull(msgId);
+         assertTrue(msgId.indexOf("STOMP") == 0);
+      }
+
+      browser.close();
+
+      MessageConsumer consumer = session.createConsumer(queue);
 
       TextMessage message = (TextMessage) consumer.receive(1000);
       Assert.assertNotNull(message);
-      Assert.assertEquals("Hello World", message.getText());
-      // Assert default priority 4 is used when priority header is not set
-      Assert.assertEquals("getJMSPriority", 4, message.getJMSPriority());
-      Assert.assertEquals("JMSXUserID", "brianm", message.getStringProperty("JMSXUserID"));
 
-      // Make sure that the timestamp is valid - should
-      // be very close to the current time.
-      long tnow = System.currentTimeMillis();
-      long tmsg = message.getJMSTimestamp();
-      Assert.assertTrue(Math.abs(tnow - tmsg) < 1000);
+      message = (TextMessage) consumer.receive(1000);
+      Assert.assertNotNull(message);
+
+      message = (TextMessage) consumer.receive(2000);
+      Assert.assertNull(message);
+
+      conn.disconnect();
    }
 }
