+++ tests/unit-tests/src/test/java/org/apache/activemq/artemis/tests/unit/core/server/group/impl/SystemPropertyOverrideTest.java
@@ -19,6 +19,7 @@ package org.apache.activemq.artemis.tests.unit.core.server.group.impl;
 import org.apache.activemq.artemis.api.core.SimpleString;
 import org.apache.activemq.artemis.core.server.group.impl.GroupingHandlerConfiguration;
 import org.apache.activemq.artemis.tests.util.ActiveMQTestBase;
+import org.junit.Test;
 
 public class SystemPropertyOverrideTest extends ActiveMQTestBase {
    // Constants -----------------------------------------------------
@@ -31,6 +32,7 @@ public class SystemPropertyOverrideTest extends ActiveMQTestBase {
 
    // Public --------------------------------------------------------
 
+   @Test
    public void testSystemPropertyOverride() throws Exception {
       final String groupTimeoutPropertyValue = "1234";
       final String reaperPeriodPropertyValue = "5678";
